hiretrack/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.7
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
phonenumbers==8.13.42
psycopg2-binary==2.9.9
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2